refactor(FE): `useToast` hook
